+++ oss/tiff/tools/fax2tiff.c
@@ -267,7 +267,7 @@ main(int argc, char* argv[])
 #if defined(_WIN32) && defined(USE_WIN32_FILEIO)
                 TIFFSetClientdata(faxTIFF, (thandle_t)_get_osfhandle(fileno(in)));
 #else
-                TIFFSetClientdata(faxTIFF, (thandle_t)fileno(in));
+                TIFFSetClientdata(faxTIFF, (thandle_t)(size_t)fileno(in));
 #endif
 		TIFFSetFileName(faxTIFF, (const char*)argv[optind]);
 		TIFFSetField(out, TIFFTAG_IMAGEWIDTH, xsize);
